make small fixes for ui and fixes for backend fixes
