+++ deploy/00crds.yaml
@@ -1140,6 +1140,7 @@ rules:
   - storagemappings
   - vjailbreaknodes
   - vmwarecreds
+  - vmwaremachines
   verbs:
   - create
   - delete
