+++ k8s/migration/config/rbac/role.yaml
@@ -65,6 +65,7 @@ rules:
   - storagemappings
   - vjailbreaknodes
   - vmwarecreds
+  - vmwaremachines
   verbs:
   - create
   - delete
