+++ k8s/migration/internal/controller/vmwarecreds_controller.go
@@ -99,6 +99,13 @@ func (r *VMwareCredsReconciler) reconcileNormal(ctx context.Context, scope *scop
 		return ctrl.Result{}, errors.Wrap(err, fmt.Sprintf("Error validating VMwareCreds '%s'", scope.Name()))
 	} else {
 		ctxlog.Info(fmt.Sprintf("Successfully authenticated to VMware '%s'", scope.Name()))
+		// Update the status of the VMwareCreds object
+		scope.VMwareCreds.Status.VMwareValidationStatus = string(corev1.PodSucceeded)
+		scope.VMwareCreds.Status.VMwareValidationMessage = "Successfully authenticated to VMware"
+		if err := r.Status().Update(ctx, scope.VMwareCreds); err != nil {
+			return ctrl.Result{}, errors.Wrap(err, fmt.Sprintf("Error updating status of VMwareCreds '%s'", scope.Name()))
+		}
+
 		vminfo, err := utils.GetAllVMs(ctx, scope.VMwareCreds, scope.VMwareCreds.Spec.DataCenter)
 		if err != nil {
 			return ctrl.Result{}, errors.Wrap(err, fmt.Sprintf("Error getting info of all VMs for VMwareCreds '%s'", scope.Name()))
@@ -107,12 +114,7 @@ func (r *VMwareCredsReconciler) reconcileNormal(ctx context.Context, scope *scop
 		if err != nil {
 			return ctrl.Result{}, errors.Wrap(err, fmt.Sprintf("Error creating VMs for VMwareCreds '%s'", scope.Name()))
 		}
-		// Update the status of the VMwareCreds object
-		scope.VMwareCreds.Status.VMwareValidationStatus = string(corev1.PodSucceeded)
-		scope.VMwareCreds.Status.VMwareValidationMessage = "Successfully authenticated to VMware"
-		if err := r.Status().Update(ctx, scope.VMwareCreds); err != nil {
-			return ctrl.Result{}, errors.Wrap(err, fmt.Sprintf("Error updating status of VMwareCreds '%s'", scope.Name()))
-		}
+
 	}
 	return ctrl.Result{RequeueAfter: constants.VMwareCredsRequeueAfter}, nil
 }
