+++ ui/src/api/openstack-creds/model.ts
@@ -63,6 +63,19 @@ export interface FSpec {
   "f:OS_USERNAME": Empty
 }
 
+export interface OpenstackFlavor {
+  id: string
+  name: string
+  vcpus: number
+  ram: number
+  disk: number
+}
+
+export interface OpenstackFlavorsResponse {
+  flavors: OpenstackFlavor[]
+}
+
+
 export interface OpenstackCredsSpec {
   // For credentials using secretRef (new approach)
   secretRef?: {
@@ -76,6 +89,7 @@ export interface OpenstackCredsSpec {
   OS_REGION_NAME?: string
   OS_TENANT_NAME?: string
   OS_INSECURE?: boolean
+  flavors?: OpenstackFlavor[]
 }
 
 export interface GetOpenstackCredsListMetadata {
